+++ app/src/index.tsx
@@ -11,6 +11,7 @@ import dashboardsPlugin from '@kobsio/plugin-dashboards';
 import prometheusPlugin from '@kobsio/plugin-prometheus';
 import elasticsearchPlugin from '@kobsio/plugin-elasticsearch';
 import jaegerPlugin from '@kobsio/plugin-jaeger';
+import markdownPlugin from '@kobsio/plugin-markdown';
 
 ReactDOM.render(
   <React.StrictMode>
@@ -22,6 +23,7 @@ ReactDOM.render(
       ...prometheusPlugin,
       ...elasticsearchPlugin,
       ...jaegerPlugin,
+      ...markdownPlugin,
     }} />
   </React.StrictMode>,
   document.getElementById('root')