+++ plugins/markdown/src/index.ts
@@ -1,0 +1,14 @@
+import { IPluginComponents } from '@kobsio/plugin-core';
+
+import icon from './assets/icon.png';
+
+import Panel from './components/panel/Panel';
+
+const markdownPlugin: IPluginComponents = {
+  markdown: {
+    icon: icon,
+    panel: Panel,
+  },
+};
+
+export default markdownPlugin;