+++ plugins/markdown/src/utils/interfaces.ts
@@ -1,0 +1,3 @@
+export interface IPanelOptions {
+  text?: string;
+}